+++ phpunit.xml.dist
@@ -21,10 +21,6 @@
         <testsuite name="Yii Input HTTP tests">
             <directory>./tests</directory>
         </testsuite>
-
-        <testsuite name="Yii Input HTTP PHP 8.1 tests">
-            <directory phpVersion="8.1" phpVersionOperator=">=">./tests/TestEnvironments/Php81</directory>
-        </testsuite>
     </testsuites>
 
     <coverage>
